fog/requirements.txt
Flask
pandas
joblib>=0.16.0
mysql-connector-python
requests>=2.24.0
numpy>=1.15.4
scikit-learn>=0.23.2